re-generated examples
