+++ examples/request reply/streetlight/README.md
@@ -1,7 +1,9 @@
-# Streetlight
 
+    # Streetlight
 
-
-
-Available nats encodings:
-https://nodejs.org/api/buffer.html#buffer_buffers_and_character_encodings
+    null
+    
+    
+    Available nats encodings:
+    https://nodejs.org/api/buffer.html#buffer_buffers_and_character_encodings
+
