+++ examples/request reply/streetlight/src/tests/testclient/testchannels/StreetlightStreetlightIdCommandTurnon.ts
@@ -1,57 +1,75 @@
 
-import * as TurnonCommandMessage from '../../../messages/TurnonCommand';
-import * as GeneralReplyMessage from '../../../messages/GeneralReply';
-import { Client, NatsError, Subscription, SubscriptionOptions, Payload } from 'ts-nats';
-import {ErrorCode, NatsTypescriptTemplateError} from '../../../NatsTypescriptTemplateError';
-import { Hooks } from '../../../hooks';
-  
-export function request(
-  message: TurnonCommandMessage.TurnonCommand,
-  nc: Client
-  
-    ,streetlight_id: string
-  
-  ): Promise<GeneralReplyMessage.GeneralReply> {
-  return new Promise(async (resolve, reject) => {
-    var timeout = undefined;
-    let msg;
-    try {
+    
+      import * as TurnonCommandMessage from '../../../messages/TurnonCommand'
       
-try {
-  let beforeSendingHooks = Hooks.getInstance().getBeforeSendingDataHook();
-  var dataToSend : any = message;
-  for(let hook of beforeSendingHooks){
-    dataToSend = hook(dataToSend);
-  }
-} catch(e){
-  const error = NatsTypescriptTemplateError.errorForCode(ErrorCode.HOOK_ERROR, e);
-  throw error;
-}
+    
+      import * as GeneralReplyMessage from '../../../messages/GeneralReply'
+      
+    import { Client, NatsError, Subscription, SubscriptionOptions, Payload } from 'ts-nats';
+    import {ErrorCode, NatsTypescriptTemplateError} from '../../../NatsTypescriptTemplateError';
+    import { Hooks } from '../../../hooks';
 
-      msg = await nc.request(`streetlight.${streetlight_id}.command.turnon`, timeout, dataToSend)
-    }catch(e){
-      reject(NatsTypescriptTemplateError.errorForCode(ErrorCode.INTERNAL_NATS_TS_ERROR, e));
-      return;
-    }
-    try{
+    
+    export function request(
+      message: TurnonCommandMessage.TurnonCommand,
+      nc: Client
       
+        ,streetlight_id: string
+        
+      ): Promise<GeneralReplyMessage.GeneralReply> {
+      return new Promise(async (resolve, reject) => {
+        var timeout = undefined;
+        undefined
+        let msg;
+        try {
+
+          
+            
+	try {
+		let beforeSendingHooks = Hooks.getInstance().getBeforeSendingDataHook();
+		var dataToSend : any = message;
+		for(let hook of beforeSendingHooks){
+		  dataToSend = hook(dataToSend);
+		}
+
+		
+
+		
+	  } catch(e) {
+		const error = NatsTypescriptTemplateError.errorForCode(ErrorCode.HOOK_ERROR, e);
+		throw error;
+	  }
+	
+            msg = await nc.request(`streetlight.${streetlight_id}.command.turnon`, timeout, dataToSend)
+            
+        }catch(e){
+          reject(NatsTypescriptTemplateError.errorForCode(ErrorCode.INTERNAL_NATS_TS_ERROR, e));
+          return;
+        }
+        
+          try{
+            
 try {
-  let receivedDataHooks = Hooks.getInstance().getreceivedDataHook();
-  var receivedData : any = msg.data;
-  for(let hook of receivedDataHooks){
-    receivedData = hook(receivedData);
-  }
+	let receivedDataHooks = Hooks.getInstance().getreceivedDataHook();
+	var receivedData : any = msg.data;
+	for(let hook of receivedDataHooks){
+		receivedData = hook(receivedData);
+	}
+	
+
+	
 } catch (e) {
-  const error = NatsTypescriptTemplateError.errorForCode(ErrorCode.HOOK_ERROR, e);
-  throw error;
+	const error = NatsTypescriptTemplateError.errorForCode(ErrorCode.HOOK_ERROR, e);
+	throw error;
 }
-
-    }catch(e){
-      reject(e)
-      return;
+	
+          }catch(e){
+            reject(e)
+            return;
+          }
+          resolve(receivedData);
+          
+      })
     }
-    resolve(receivedData);
-  })
-}
-
-
+    
+
